social-flow-backend/testing/integration/messaging-tests/requirements.txt
pytest>=8.0
pytest-asyncio>=0.21
aiokafka>=0.8.0          # async Kafka client
kafka-python>=2.0.2      # sync Kafka client (fallback)
pika>=1.3.1              # RabbitMQ client
aioredis>=2.0.0          # Redis async client
redis>=4.5.0             # sync Redis client
jsonschema>=4.17
fastavro>=1.7.0          # optional Avro validation
tenacity>=8.2
python-dotenv>=1.0
pytest-xdist>=3.2
locust>=2.20             # optional perf
